% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data Points from Kolada}
\usage{
get_data(kpi_id = NULL, municipality_id = NULL, year = NULL)
}
\arguments{
\item{kpi_id}{A character vector of one or more KPI IDs (e.g., "N01951").}

\item{municipality_id}{A character vector of one or more municipality IDs
(e.g., "0580").}

\item{year}{A character or numeric vector of one or more years
(e.g., 2022 or "2022").}
}
\value{
A tibble with the requested data.
}
\description{
Fetches specific data points from the Kolada API based on a combination of
query conditions. The function is flexible and can handle
single values or vectors for each parameter.
}
\examples{
\dontrun{
# Get total population for Linköping (0580) for the year 2022
population_data <- get_data(kpi_id = "N01951", municipality_id = "0580", year = "2022")

# Get data for multiple KPIs and multiple years for multiple municipalities
multi_data <- get_data(kpi_id = c("N01951", "N01953"),
                       municipality_id = c("0580", "0001"),
                       year = c("2020", "2021", "2022"))
}
}
