% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpis.R
\name{get_kpis}
\alias{get_kpis}
\title{Get Key Performance Indicators (KPIs)}
\usage{
get_kpis(title = NULL)
}
\arguments{
\item{title}{A character string used to search for KPIs by title.
The search uses a space-separated list of filter arguments. Can be NULL.#'}
}
\value{
A tibble with columns `id`, `title`, and `description`.
}
\description{
Fetches a list of Key Performance Indicators (KPIs) from the Kolada API.
You can retrieve all KPIs or filter them by providing a search string for
the title.
}
\examples{
\dontrun{
# Get all KPIs
all_kpis <- get_kpis()
# Search for KPIs with "Invånare totalt, antal" (Total inhabitants, number) in the title
population_kpis <- get_kpis(title = "Invånare totalt, antal")
}
}
