% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get Municipalities}
\usage{
get_municipalities(title = NULL)
}
\arguments{
\item{title}{A character string used to search for municipalities by title.
The search uses a space-separated list of filter arguments. Can be NULL.}
}
\value{
A tibble with columns `id` and `title`.
}
\description{
Fetches a list of Swedish municipalities from the Kolada API.
You can retrieve all municipalities or filter them by providing a search
string for the title.
}
\examples{
\dontrun{
# Get all municipalities
all_municipalities <- get_municipalities()

# Search for municipalities with "Linköping" in the title
linkoping <- get_municipalities(title = "Linköping")
}
}
